Update jquery.ui.touch-punch.js
Allow a little movement before deciding that a touch is not a click (coded to allow 5 move events - but could be a parameter).

This allows sensitive screens (like nexus10) to better register clicks - without this, it is virtually impossible for a user to 'click' as there is always slight movement between touch and release.

The downside is that 'flicking' an item to move it slightly, may register as a click also. The sensitivity could be increased (allow fewer moves between touch and release), or parameterised to make it easily for application users to adjust. Or a genuine move could be de-bounced (if a move happens, any potential click is cancelled).