Drag and Drop in another iframe
When the drag and drop is made in another iframe, like the children iframe, the jQuery UI draggable widget add the mousemove and mouseup listener on the document element of the page.
Dispatching the event to the dragging element that is in the iframe is not enought to make the event to reach the parent document,
the event remain in the iframe. So I have added the dispatch of the event to the document too.